Set up the search form with livesearch and add new component submissionForm
