+++ src/Components/App/App.tsx
@@ -2,10 +2,10 @@ import React, { useEffect, useState } from 'react';
 // import { Route } from 
 import './App.css';
 import Nav from '../../Components/Nav/Nav'
-import Form from '../Form/Form';
 import { User } from '../../interfaces'
 import { user1 } from '../../mockData'
 import { Route } from 'react-router-dom'
+import SearchForm from '../Form/SearchForm';
 
 const App = () => {
   const [user, setUser] = useState<User>()
@@ -19,12 +19,9 @@ const App = () => {
     <div className="App">
       {user ? <Nav name={user.name} /> : <Nav name={'No user found'} />}
       <Route exact path='/add-new'>
-        {user ? <Form userID={user.id} /> : <Form userID={0} />}
-
+        {user ? <SearchForm userID={user.id} /> : <SearchForm userID={0} />}
       </Route>
-
     </div>
-
   );
 }
