+++ src/Components/Form/SearchForm.css
@@ -3,6 +3,7 @@
     width: 40vw;
     height: 30vh;
     margin-left: 30vw;
+    overflow-y: auto;
 }
 
 .form__search-result:hover{
