+++ src/Components/Form/SubmissionForm.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+type MedProps = {
+    chosenMedicine: string
+}
+
+const SubmissionForm: React.FC<MedProps> = ({chosenMedicine}) => {
+    return (
+        <>
+            <p>Submission Form Starts Here</p>
+        </>
+    )
+}
+
+
+
+export default SubmissionForm
